feat:NavigationCountView로 분리
